Backend/requirements.txt
# Civilization Shield - Django Backend
# Python 3.9+

# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# External API calls
requests==2.31.0

# Caching
cachetools==5.3.2

# Environment variables
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
geopy==2.4.1
